fix: 修复费用聚合整数溢出

+++ src-tauri/src/infra/request_logs/types.rs
@@ -157,6 +157,6 @@ pub struct SessionStatsAggregate {
     pub request_count: i64,
     pub total_input_tokens: i64,
     pub total_output_tokens: i64,
-    pub total_cost_usd_femto: i64,
+    pub total_cost_usd_femto: f64,
     pub total_duration_ms: i64,
 }
